+++ packages/internal/babel-compile/package.json
@@ -1,0 +1,10 @@
+{
+  "name": "@internal/babel-compile",
+  "private": true,
+  "version": "0.19.2",
+  "bin": "./bin/index.js",
+  "dependencies": {
+    "@babel/cli": "^7.0.0",
+    "@internal/helpers": "0.19.2"
+  }
+}
