+++ packages/internal/helpers/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "@internal/helpers",
+  "private": true,
+  "version": "0.19.2",
+  "bin": "./bin/index.js",
+  "main": "./index.js",
+  "dependencies": {
+    "glob": "7.1.3",
+    "minimist": "1.2.0"
+  }
+}
