+++ packages/internal/helpers/safe-name.js
@@ -1,0 +1,3 @@
+const safeName = name => `${name.replace(/@/, '').replace(/[/|\\]/g, '-')}.tgz`
+
+module.exports = { safeName }
