+++ packages/internal/local-registry/package.json
@@ -1,0 +1,9 @@
+{
+  "name": "@internal/local-registry",
+  "private": true,
+  "version": "0.19.2",
+  "bin": "./bin/index.js",
+  "dependencies": {
+    "@internal/helpers": "0.19.2"
+  }
+}
